Fixed a bunch of compiler warnings on Windows.

+++ gdk/gdk_utils.c
@@ -747,7 +747,7 @@ GDKfree_(void *blk)
 	DEADBEEFCHK memset(s, 0xDB, size - (MALLOC_EXTRA_SPACE + (size & 1)));	/* 0xDeadBeef */
 #endif
 	free(((char *) s) - MALLOC_EXTRA_SPACE);
-	heapdec((size_t) size);
+	heapdec(size);
 }
 
 void
@@ -810,7 +810,7 @@ GDKreallocmax(void *blk, size_t size, size_t *maxsize, int emergency)
 
 	/* adapt statistics */
 	heapinc(newsize);
-	heapdec((size_t) oldsize);
+	heapdec(oldsize);
 	*maxsize = size;
 	return blk;
 }
